feat(habitat): move pattern governance into domains
Move diagnostic catalog ownership from src/lib into src/domains/diagnostic-pattern-catalog and move pattern admission, manifest validation, states, and views from src/rules/patterns into src/domains/pattern-governance. Update callers and public exports to use the domain ownership paths directly.

Keep the Grit provider below domain ownership by moving the diagnostic acquisition parser/projection from providers/grit/output into the Grit adapter. The provider now owns raw Grit wire/result types and options; the adapter combines provider output with Habitat diagnostic-domain observations.

Add a service architecture guard that fails if active diagnostic catalog or pattern governance imports drift back to src/lib/diagnostic-catalog or src/rules/patterns.

Validation: bun run --cwd tools/habitat-harness check; bun run --cwd tools/habitat-harness test -- test/lib/grit-adapter.test.ts test/rules/pattern-manifest.test.ts test/rules/pattern-views.test.ts test/generators/pattern-generator.test.ts test/lib/pattern-apply.test.ts test/service/service-architecture.test.ts test/service/fix-service.test.ts test/service/transactions-service.test.ts; bun run --cwd tools/habitat-harness validate:grit-patterns; bun run biome:ci; bun run openspec -- validate deep-habitat-effect-diagnostic-pattern-governance --strict; bun run openspec:validate; git diff --check; bun run build.

Note: bun run build completed successfully and reported existing Nx flaky-task metadata for @civ7/adapter:build.
