+++ tools/habitat-harness/src/adapters/grit/diagnostics.ts
@@ -6,7 +6,7 @@ import {
   observedGritDiagnosticIdentity,
   observedGritIdentityMatches,
   renderUnexpectedObservedGritIdentity,
-} from "../../lib/diagnostic-catalog/index.js";
+} from "../../domains/diagnostic-pattern-catalog/index.js";
 import type { GritDiagnosticOptions, GritReport, GritResult } from "../../providers/grit/types.js";
 import type { RuleRunResult } from "../../rules/architecture.js";
 import type { RulePatternFacts } from "../../rules/registry/index.js";
