+++ tools/habitat-harness/src/adapters/grit/failure.ts
@@ -1,7 +1,7 @@
 import {
   type DiagnosticAdapterFailureKind,
   renderDiagnosticAdapterFailure,
-} from "../../lib/diagnostic-catalog/index.js";
+} from "../../domains/diagnostic-pattern-catalog/index.js";
 import type { RuleRunResult } from "../../rules/architecture.js";
 import type { RulePatternFacts } from "../../rules/registry/index.js";
