+++ tools/habitat-harness/src/domains/diagnostic-pattern-catalog/catalog.ts
