+++ tools/habitat-harness/src/domains/diagnostic-pattern-catalog/command.ts
