+++ tools/habitat-harness/src/domains/diagnostic-pattern-catalog/failure.ts
