+++ tools/habitat-harness/src/domains/diagnostic-pattern-catalog/identity.ts
