+++ tools/habitat-harness/src/domains/diagnostic-pattern-catalog/index.ts
