+++ tools/habitat-harness/src/domains/diagnostic-pattern-catalog/outcome.ts
