+++ tools/habitat-harness/src/domains/pattern-governance/apply-admissions.ts
@@ -1,6 +1,6 @@
 import { Value } from "typebox/value";
-import { activeRulePatternFacts } from "../facts.js";
-import type { RulePatternFacts } from "../registry/index.js";
+import { activeRulePatternFacts } from "../../rules/facts.js";
+import type { RulePatternFacts } from "../../rules/registry/index.js";
 import { applyAdmittedState } from "./admission.js";
 import {
   type ApplyAdmission,
