+++ tools/habitat-harness/src/domains/pattern-governance/paths.ts
