+++ tools/habitat-harness/src/domains/pattern-governance/rule-reference.ts
