+++ tools/habitat-harness/src/domains/pattern-governance/schema.ts
