+++ tools/habitat-harness/src/domains/pattern-governance/validation.ts
