+++ tools/habitat-harness/src/lib/pattern-apply/transaction-input.ts
@@ -6,7 +6,7 @@ import {
   ApplyDryRunCommandSchema,
   type ApplyTransactionInput,
   ApplyTransactionInputSchema,
-} from "../../rules/patterns/index.js";
+} from "../../domains/pattern-governance/index.js";
 import { NonEmptyStringSchema } from "./primitives.js";
 
 export const GritDryRunCommandInputSchema = ApplyDryRunCommandSchema;
