+++ tools/habitat-harness/src/service/modules/transactions/context.ts
@@ -1,6 +1,6 @@
 import type { Layer } from "effect";
+import type { ApplyTransactionInput } from "../../../domains/pattern-governance/index.js";
 import type { GritProvider } from "../../../providers/grit/index.js";
-import type { ApplyTransactionInput } from "../../../rules/patterns/index.js";
 
 export interface TransactionsServiceContext {
   readonly transactions?: TransactionsServiceOptions;
