+++ tools/habitat-harness/test/lib/pattern-apply.test.ts
@@ -1,6 +1,10 @@
 import { Effect } from "effect";
 import { Value } from "typebox/value";
 import { describe, expect, test } from "vitest";
+import type {
+  ApplyAdmission,
+  ApplyTransactionInput,
+} from "../../src/domains/pattern-governance/index.js";
 import {
   type PatternApplyRequest,
   PatternApplyRequestSchema,
@@ -11,7 +15,6 @@ import {
   makeHabitatCommandResult,
 } from "../../src/providers/command/index.js";
 import { makeFakeGritProviderLayer } from "../../src/providers/grit/index.js";
-import type { ApplyAdmission, ApplyTransactionInput } from "../../src/rules/patterns/index.js";
 import type { TransactionsServiceOptions } from "../../src/service/modules/transactions/context.js";
 import { runTransactionApplyService } from "../../src/service/modules/transactions/router.js";
