+++ tools/habitat-harness/test/service/fix-service.test.ts
@@ -1,11 +1,14 @@
 import { Effect } from "effect";
 import { describe, expect, test } from "vitest";
+import type {
+  ApplyAdmission,
+  ApplyTransactionInput,
+} from "../../src/domains/pattern-governance/index.js";
 import {
   type HabitatProcessRequest,
   makeHabitatCommandResult,
 } from "../../src/providers/command/index.js";
 import { makeFakeGritProviderLayer } from "../../src/providers/grit/index.js";
-import type { ApplyAdmission, ApplyTransactionInput } from "../../src/rules/patterns/index.js";
 import { createHabitatServiceClient } from "../../src/service/client.js";
 import { runFixService } from "../../src/service/modules/fix/router.js";
